* Need to figure out how to check if record needs to be deleted.

+++ processor/app.py
@@ -46,11 +46,15 @@ def add_grids():
     # Add to list.
     logging.debug(f"Adding {solved_string} to list")
     grid_strings.append("Initial Grid: <br>" + initial_string + "<br>Solved Grid: <br>" + solved_string)
-    logging.debug("Moving file")
     # Adding record to database
     requests.post("http://sudoku_database:3000/database", json={"Result": 'y'})
-    # TODO: Find way not to delete first record + find length of database.
-    requests.delete("http://sudoku_database:3000/database")
+    logging.debug("Added to database")
+    # Checking if we are the first record in the database.
+    is_only_one = requests.get("http://sudoku_database:3000/item")
+    # If not, delete previous record.
+    if is_only_one.json() == json.dumps('False'):
+        requests.delete("http://sudoku_database:3000/database")
+        logging.debug("record deleted")
     return '', 204
